+++ applications/fishing-map/src/data/default-workspaces/workspace.production.ts
@@ -45,7 +45,7 @@ const workspace: Workspace<WorkspaceState> = {
       // update it here if you want to show it again or go to
       // applications/fishing-map/src/features/workspace/highlight-panel/highlight-panel.content.ts
       // and update the `dataviewInstanceId`
-      id: 'highlight-vms-with-ecuador',
+      id: 'fishing-vms',
       config: {
         color: '#FFAA0D',
         colorRamp: 'orange',
@@ -54,7 +54,6 @@ const workspace: Workspace<WorkspaceState> = {
           'public-indonesia-fishing-effort:v20200320',
           'public-panama-fishing-effort:v20200331',
           'public-peru-fishing-effort:v20200324',
-          'public-ecuador-fishing-effort:v20210612',
         ],
       },
       dataviewId: DEFAULT_FISHING_DATAVIEW_ID,