+++ applications/fishing-map/src/features/search/Search.tsx
@@ -356,6 +356,7 @@ function Search() {
                         callsign,
                         geartype,
                         origin,
+                        casco,
                         matricula,
                         dataset,
                         firstTransmissionDate,
@@ -409,6 +410,18 @@ function Search() {
                                     : EMPTY_FIELD_PLACEHOLDER}
                                 </span>
                               </div>
+                              {matricula && (
+                                <div className={styles.property}>
+                                  <label>{t('vessel.matricula', 'Matricula')}</label>
+                                  <span>{matricula}</span>
+                                </div>
+                              )}
+                              {casco && (
+                                <div className={styles.property}>
+                                  <label>{t('vessel.casco', 'Casco')}</label>
+                                  <span>{casco}</span>
+                                </div>
+                              )}
                               {fleet && (
                                 <div className={styles.property}>
                                   <label>{t('vessel.fleet', 'Fleet')}</label>
@@ -419,12 +432,6 @@ function Search() {
                                 <div className={styles.property}>
                                   <label>{t('vessel.origin', 'Origin')}</label>
                                   <span>{formatInfoField(origin, 'fleet')}</span>
-                                </div>
-                              )}
-                              {matricula && (
-                                <div className={styles.property}>
-                                  <label>{t('vessel.matricula', 'Matricula')}</label>
-                                  <span>{formatInfoField(matricula, 'fleet')}</span>
                                 </div>
                               )}
                               {firstTransmissionDate && lastTransmissionDate && (